requirements.txt
requests>=2.31.0,<3.0.0
python-frontmatter>=1.1.0,<2.0.0
minsearch>=0.1.0
sentence-transformers>=3.0.0,<4.0.0
pydantic-ai>=0.1.0
openai>=1.40.0,<2.0.0
pandas>=2.2.0,<3.0.0
streamlit>=1.38.0,<2.0.0
numpy>=1.26.0,<3.0.0
